% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGPL}
\alias{addGPL}
\title{Add GEO Platform to Operend}
\usage{
addGPL(GPL)
}
\arguments{
\item{GPL}{A character vector of length 1 specifying a GEO Platform accession number, OR
a GEOquery GEO Platform object.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Platform.
}
\description{
This function queries Operend for the accession number of a GEO Platform. If 
the GPL has previously been uploaded, retrieves the \code{Entity} record and
checks for any updates. If the GPL has not been uploaded, extracts and
uploads metadata from a GEO Platform as a new Operend \code{Entity} record.
}
\examples{
GPL <- getGEO("GPL96", GSEMatrix = FALSE)
addGPL(GPL)
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
