% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGSE}
\alias{addGSE}
\title{Add GEO Series to Operend}
\usage{
addGSE(GSE)
}
\arguments{
\item{GSE}{A character vector of length 1 specifying a GEO Series accession number OR
a GEOquery GEO Series object.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Series.
}
\description{
This function queries Operend for the accession number of a GEO Series. If 
the GSE has previously been uploaded, retrieves the \code{Entity} record and
checks for any updates. If the GSE has not been uploaded, extracts and
uploads metadata from a GEO Series and associated Platforms, Samples, and 
affymetrixCEL files as a new Operend \code{Entity} record.
}
\examples{
addGSE("GSE994")
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
