% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{geoPermissions}
\alias{geoPermissions}
\title{Permissions for GEO Operations}
\format{
A list with the following components:
\describe{
  \item{geo}{A character vector with permissions "R", "U", "PR", "PU".}
  \item{_other}{A character vector with permission "R".}
}
}
\usage{
geoPermissions
}
\description{
This object contains permission settings for adding GSE, GPL, GSM,
affymetrixCEL, and affymetrixCELSet Entities, as well as associated work files.
}
\examples{
opeRend.NCBI::geoPermissions
}
\seealso{
\code{\link[opeRend]{operendPermissions}}
}
\keyword{datasets}
