% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCELurl}
\alias{getCELurl}
\title{Return first matching CEL file ftp link}
\usage{
getCELurl(supplementaryFile)
}
\arguments{
\item{supplementaryFile}{A character vector or \code{NULL}.}
}
\value{
A character vector of length 1 representing a ftp link to a CEL file,
OR \code{NULL}.
}
\description{
Helper function that checks character vector for an ftp link to a CEL file.
}
