% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSubseries}
\alias{getSubseries}
\title{Return GSE accession numbers of subseries}
\usage{
getSubseries(relationVector)
}
\arguments{
\item{relationVector}{A character vector corresponding to the relation metadata of a GSE}
}
\value{
A character vector representing matches for GSE subseries, OR \code{NULL}.
}
\description{
Helper function that extracts
}
