% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{needsUpdate}
\alias{needsUpdate}
\title{Logic for comparing last updated dates.}
\usage{
needsUpdate(entity, last_update_date)
}
\arguments{
\item{entity}{An Operend GEO entity containing a last_update_date variable.}

\item{last_update_date}{A character vector of length 1 representing an operendDate.}
}
\value{
A character vector of length 1 representing a formatted operendDate.
}
\description{
Helper function that returns TRUE if two dates are not equal (therefore
requiring an update), or FALSE if two dates are equal (therefore indicating
that the provided Operend entity is up to date with the NCBI entry).
}
