% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processCELHeader.R
\name{Process CEL header}
\alias{Process CEL header}
\alias{processCELHeader}
\title{Process the header of a CEL file}
\usage{
processCELHeader(filename)
}
\arguments{
\item{filename}{A character vector of length 1 containing the full path to a CEL file}
}
\value{
This function returns a named list containing the following elements:
\describe{
  \item{cdfName}{
    A character string containing the \code{cdfName} field of the header
  }
  \item{dimensions}{
    A two-element, unnamed integer vector containing the
    \code{CEL dimensions} field of the header
  }
  \item{DatHeader}{
    A character string containing the \code{DatHeader} field of the header
  }
  \item{ScanDate}{
    A character string containing the \code{ScanDate} field of the header,
    with extraneous characters removed
  }
}
}
\description{
This function extracts and processes the header of a CEL file.
}
\author{
Adam C. Gower \email{agower@bu.edu}
}
