% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processDate}
\alias{processDate}
\title{Format dates for Operend date variables.}
\usage{
processDate(d)
}
\arguments{
\item{d}{A character vector of length 1 representing a date.}
}
\value{
A character vector of length 1 representing a formatted operendDate.
}
\description{
Format dates for Operend date variables.
}
\examples{
processDate("Feb 19 2004")
}
