% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{queryOperend}
\alias{queryOperend}
\title{Query Operend for GEO accession number}
\usage{
queryOperend(class, variables)
}
\arguments{
\item{class}{A character vector of length 1 specifying the Operend \code{EntityClass}
of a GEO entity.}

\item{variables}{A named list indicating variables in the Operend Entity to search for.}
}
\value{
An \code{\linkS4class{operendEntity}} object, or \code{NULL}.
}
\description{
Query Operend for the GEO accession number of a GEO entity, returning the
least recently added entry.
}
