% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{removeNull}
\alias{removeNull}
\title{Remove \code{NULL} values from a list}
\usage{
removeNull(list_obj)
}
\arguments{
\item{list_obj}{A list.}
}
\value{
A list with all \code{NULL} values removed.
}
\description{
Remove \code{NULL} values from a list
}
