% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retrieveGEOquery}
\alias{retrieveGEOquery}
\title{Retrieve GEOquery object based on th}
\usage{
retrieveGEOquery(GEO, GEOqueryClass)
}
\arguments{
\item{GEO}{A character vector of length 1 specifying a GEO accession number,
or a GEOquery GEO object.}

\item{GEOqueryClass}{A character vector of length 1 specifying the class of
the GEOquery GEO object to be returned.}
}
\value{
A GEOquery GEO object.
}
\description{
Retrieve GEOquery object based on th
}
