% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addAffymetrixCEL}
\alias{addAffymetrixCEL}
\title{Add affymetrixCEL file to Operend}
\usage{
addAffymetrixCEL(ftpUrl)
}
\arguments{
\item{ftpUrl}{Character vector of length 1 specifying the ftp link of the CEL file of a GEO Sample.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the affymetrixCEL.
}
\description{
This function extracts and uploads the metadata and workfile of an affymetrixCEL
file as an Operend \code{Entity} record and \code{WorkFile}.
}
\examples{
addAffymetrixCEL(
  "ftp://ftp.ncbi.nlm.nih.gov/geo/samples/GSM15nnn/GSM15684/suppl/GSM15684.CEL.gz"
)
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
