% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGPL}
\alias{addGPL}
\title{Add GEO Platform to Operend}
\usage{
addGPL(GPL)
}
\arguments{
\item{GPL}{A GEOquery GEO Platform entity.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Platform.
}
\description{
This function queries Operend for the accession number of a GEO Platform; if 
the GPL has previously been uploaded, retrieves the \code{Entity} record,
otherwise extracts and uploads metadata from a GEO Platform as a new Operend
\code{Entity} record.
}
\examples{
GSE <- getGEO("GSE994", GSEMatrix = FALSE)
lapply(GPLList(GSE), addGPL)
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
