% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGSE}
\alias{addGSE}
\title{Add GEO Series to Operend}
\usage{
addGSE(GSE)
}
\arguments{
\item{GSE}{A character vector of length 1 specifying a GEO Series accession number.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Series.
}
\description{
This function extracts and uploads metadata from a GEO Series, including associated
Platforms, Samples, and affymetrixCEL files, as an Operend \code{Entity} record.
}
\examples{
addGSE("GSE994")
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
