% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGSM}
\alias{addGSM}
\title{Add GEO Sample to Operend}
\usage{
addGSM(GSM)
}
\arguments{
\item{GSM}{A GEOquery GEO Sample entity.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Sample.
}
\description{
This function extracts and uploads metadata from a GEO Sample, including associated
affymetrixCEL files, as an Operend \code{Entity} record.
}
\examples{
GSE <- getGEO("GSE994", GSEMatrix = FALSE)
lapply(GSMList(GSE)[1:2], addGPL)
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
