% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEOEntities.R
\name{addGSM}
\alias{addGSM}
\title{Add GEO Sample to Operend}
\usage{
addGSM(GSM)
}
\arguments{
\item{GSM}{A character vector of length 1 specifying a GEO Sample accession number, OR
a GEOquery GEO Sample object.}
}
\value{
If the operation is successful, an \code{\linkS4class{operendEntity}} object of the GEO Sample.
}
\description{
This function queries Operend for the accession number of a GEO Sample. If 
the GSM has previously been uploaded, retrieves the \code{Entity} record and
checks for any updates. If the GSM has not been uploaded, extracts and
uploads metadata from a GEO Sample and associated affymetrixCEL files as a
new Operend \code{Entity} record.
}
\examples{
GSE <- GEOquery::getGEO("GSE994", GSEMatrix = FALSE)
lapply(GSMList(GSE)[1:2], addGPL)
}
\author{
Dylan L. Tamayo \email{dltamayo@bu.edu}
}
